add mapfilter-db as dependency and update features conversion
