+++ package.json
@@ -69,7 +69,6 @@
   "dependencies": {
     "blob-to-buffer": "^1.2.6",
     "clone": "^2.1.1",
-    "collect-stream": "^1.2.1",
     "drag-drop": "^2.13.2",
     "ecstatic": "^2.1.0",
     "electron-compile": "^6.4.2",
@@ -81,20 +80,16 @@
     "from2": "^2.3.0",
     "get-ports": "^1.0.3",
     "glob": "^7.1.2",
-    "hyperlog-sneakernet-replicator": "^1.1.4",
     "imghdr": "^0.1.6",
     "jpeg-marker-stream": "^1.0.0",
-    "level": "^1.6.0",
     "lodash": "^4.17.4",
     "material-ui": "^1.0.0-beta.35",
     "material-ui-icons": "^1.0.0-beta.35",
+    "mapfilter-db": "^1.0.1",
     "mime": "^1.3.4",
     "mkdirp": "^0.5.1",
     "object-assign": "^4.1.1",
     "once": "^1.4.0",
-    "osm-p2p": "^1.4.0",
-    "osm-p2p-observations": "^2.0.0",
-    "p2p-file-store": "^0.0.1",
     "prop-types": "^15.6.0",
     "pump": "^1.0.2",
     "randombytes": "^2.0.3",
